requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0
unidiff>=0.7.0
sqlalchemy>=2.0.0
celery>=5.3.0
redis>=5.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.16.0
pyyaml>=6.0
